move exceptions to their own folder
