+++ src/main/java/org/corfudb/runtime/exceptions/HoleEncounteredException.java
@@ -1,4 +1,4 @@
-package org.corfudb.runtime;
+package org.corfudb.runtime.exceptions;
 
 import lombok.Getter;
 import org.corfudb.runtime.stream.ITimestamp;
