+++ src/main/java/org/corfudb/runtime/exceptions/LinearizationException.java
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package org.corfudb.runtime;
+package org.corfudb.runtime.exceptions;
 
 import org.corfudb.runtime.stream.Timestamp;
