+++ src/main/java/org/corfudb/runtime/exceptions/NetworkException.java
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package org.corfudb.runtime;
+package org.corfudb.runtime.exceptions;
 import java.io.IOException;
 import org.corfudb.runtime.protocols.IServerProtocol;
 /**
