+++ src/main/java/org/corfudb/runtime/exceptions/NonStreamEntryException.java
@@ -1,4 +1,4 @@
-package org.corfudb.runtime;
+package org.corfudb.runtime.exceptions;
 
 import java.io.IOException;
