+++ src/main/java/org/corfudb/runtime/exceptions/TransactionAbortedException.java
@@ -1,4 +1,4 @@
-package org.corfudb.runtime;
+package org.corfudb.runtime.exceptions;
 
 /**
  * Created by mwei on 10/2/15.
