+++ src/main/java/org/corfudb/runtime/exceptions/UnwrittenException.java
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package org.corfudb.runtime;
+package org.corfudb.runtime.exceptions;
 import java.io.IOException;
 
 /**
