+++ src/main/java/org/corfudb/runtime/protocols/configmasters/IConfigMaster.java
@@ -14,7 +14,7 @@
  */
 
 package org.corfudb.runtime.protocols.configmasters;
-import org.corfudb.runtime.NetworkException;
+import org.corfudb.runtime.exceptions.NetworkException;
 import org.corfudb.runtime.protocols.IServerProtocol;
 
 import org.corfudb.runtime.view.CorfuDBView;
