+++ src/main/java/org/corfudb/runtime/protocols/replications/ChainReplicationProtocol.java
@@ -2,11 +2,10 @@ package org.corfudb.runtime.protocols.replications;
 
 import org.corfudb.infrastructure.thrift.Hints;
 import org.corfudb.runtime.*;
+import org.corfudb.runtime.exceptions.*;
 import org.corfudb.runtime.protocols.IServerProtocol;
 import org.corfudb.runtime.protocols.logunits.IWriteOnceLogUnit;
 import org.corfudb.runtime.smr.MultiCommand;
-import org.corfudb.runtime.view.CorfuDBView;
-import org.corfudb.runtime.view.CorfuDBViewSegment;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
