+++ src/main/java/org/corfudb/runtime/protocols/replications/IReplicationProtocol.java
@@ -2,9 +2,9 @@ package org.corfudb.runtime.protocols.replications;
 
 import org.corfudb.infrastructure.thrift.Hints;
 import org.corfudb.runtime.*;
+import org.corfudb.runtime.exceptions.*;
 import org.corfudb.runtime.protocols.IServerProtocol;
 import org.corfudb.runtime.smr.MultiCommand;
-import org.corfudb.runtime.view.CorfuDBView;
 
 import javax.json.JsonObject;
 import java.io.IOException;
