+++ src/main/java/org/corfudb/runtime/protocols/sequencers/IStreamSequencer.java
@@ -14,7 +14,7 @@
  */
 
 package org.corfudb.runtime.protocols.sequencers;
-import org.corfudb.runtime.NetworkException;
+import org.corfudb.runtime.exceptions.NetworkException;
 import org.corfudb.runtime.protocols.IServerProtocol;
 
 import java.util.UUID;
