+++ src/main/java/org/corfudb/runtime/protocols/sequencers/RedisSequencerProtocol.java
@@ -16,7 +16,7 @@
 package org.corfudb.runtime.protocols.sequencers;
 
 import org.corfudb.runtime.protocols.IServerProtocol;
-import org.corfudb.runtime.NetworkException;
+import org.corfudb.runtime.exceptions.NetworkException;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
