+++ src/main/java/org/corfudb/runtime/smr/BufferedSMREngine.java
@@ -2,8 +2,7 @@ package org.corfudb.runtime.smr;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.corfudb.runtime.CorfuDBRuntime;
-import org.corfudb.runtime.HoleEncounteredException;
+import org.corfudb.runtime.exceptions.HoleEncounteredException;
 import org.corfudb.runtime.entries.IStreamEntry;
 import org.corfudb.runtime.smr.smrprotocol.SMRCommand;
 import org.corfudb.runtime.stream.IStream;
