+++ src/main/java/org/corfudb/runtime/view/ConfigurationMaster.java
@@ -1,10 +1,9 @@
 package org.corfudb.runtime.view;
 
 import org.corfudb.runtime.CorfuDBRuntime;
-import org.corfudb.runtime.NetworkException;
+import org.corfudb.runtime.exceptions.NetworkException;
 import org.corfudb.runtime.protocols.IServerProtocol;
 import org.corfudb.runtime.protocols.configmasters.IConfigMaster;
-import org.corfudb.runtime.protocols.sequencers.ISimpleSequencer;
 import org.corfudb.util.retry.ExponentialBackoffRetry;
 import org.corfudb.util.retry.IRetry;
 import org.slf4j.Logger;
