+++ src/main/java/org/corfudb/runtime/view/StreamAddressSpace.java
@@ -2,23 +2,13 @@ package org.corfudb.runtime.view;
 
 import com.github.benmanes.caffeine.cache.AsyncLoadingCache;
 import com.github.benmanes.caffeine.cache.Caffeine;
-import com.github.benmanes.caffeine.cache.LoadingCache;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
-import lombok.val;
 import org.corfudb.infrastructure.thrift.ReadCode;
-import org.corfudb.runtime.NetworkException;
-import org.corfudb.runtime.OutOfSpaceException;
-import org.corfudb.runtime.OverwriteException;
-import org.corfudb.runtime.TrimmedException;
 import org.corfudb.runtime.protocols.IServerProtocol;
 import org.corfudb.runtime.protocols.logunits.INewWriteOnceLogUnit;
-import org.corfudb.util.retry.ExponentialBackoffRetry;
-import org.corfudb.util.retry.IRetry;
 
-import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.util.*;
 import java.util.concurrent.CompletableFuture;
