+++ src/main/java/org/corfudb/runtime/view/WriteOnceAddressSpace.java
@@ -17,12 +17,9 @@ package org.corfudb.runtime.view;
 
 import org.corfudb.infrastructure.thrift.Hints;
 import org.corfudb.runtime.*;
-import org.corfudb.runtime.protocols.IServerProtocol;
-import org.corfudb.runtime.protocols.logunits.IWriteOnceLogUnit;
+import org.corfudb.runtime.exceptions.*;
 import org.corfudb.runtime.protocols.replications.IReplicationProtocol;
 import org.corfudb.runtime.smr.MultiCommand;
-import org.corfudb.runtime.smr.Pair;
-import org.corfudb.util.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
