+++ src/test/java/org/corfudb/runtime/view/WriteOnceAddressSpaceTest.java
@@ -1,11 +1,6 @@
 package org.corfudb.runtime.view;
 
 import org.corfudb.runtime.CorfuDBRuntime;
-import org.corfudb.runtime.OverwriteException;
-import org.corfudb.runtime.UnwrittenException;
-import org.corfudb.runtime.smr.Pair;
-import org.junit.Before;
-import org.junit.Test;
 
 import static com.github.marschall.junitlambda.LambdaAssert.assertRaises;
 import static org.junit.Assert.assertEquals;
